requirements.txt
jsonpickle==3.0.2
pandas~=2.0.3
geopandas~=0.13.2
matplotlib~=3.8.1
setuptools==69.0.2
python-dwca-reader~=0.16.0
pylint==3.0.2
autopep8==2.0.4
jsonpickle==3.0.2